+++ docs/FR1.0/test.md
@@ -1,0 +1,11 @@
+---
+title: 简单样例
+order: 1
+nav:
+  order: 5
+  title: FR 1.0
+# sidemenu: false
+gapless: true
+---
+
+<code src='./index.jsx'  />
