+++ packages/FR1.0/.fatherrc.js
@@ -1,14 +1,7 @@
-import commonjs from 'rollup-plugin-commonjs';
-
 export default {
   entry: ['src/index.js'],
   esm: 'rollup',
-  cjs: 'babel',
-  extraRollupPlugins: [
-    commonjs({
-      include: 'node_modules/**',
-    }),
-  ],
+  cjs: 'rollup',
   extraBabelPlugins: [
     [
       'import',
