+++ packages/FR1.0/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "new-form-render",
+  "name": "form-render-beta",
   "version": "1.0.0",
   "description": "通过 JSON Schema 生成标准 Form，常用于自定义搭建配置界面生成",
   "repository": {
