+++ packages/FR1.0/src/FR/RenderChildren/RenderObject.js
@@ -2,7 +2,7 @@ import React from 'react';
 import FR from '../index';
 
 // TODO: nanoId 好像没啥用
-const RenderObject = ({ children = [], dataIndex = [] }) => {
+const RenderObject = ({ children = [], dataIndex = [], errorFields }) => {
   return (
     <>
       {children.map((child, i) => {
